Splitting reward functions and fixing error
The ChromophoreReward worked, but was a bit tough to understand in the output. So, I've split it into three distinct reward functions, one for each property. As well, making this change elucidated that some of the code was not set up properly to handle multiple reward functions, so I've corrected those issues. Another file upload to follow.